Initial Commit
Currently working on saving the preprocessed images
to disk to be used for training.

There is a bug in GlaucomaDataset.save method where importing
the class in main.py/save.py results an infinate loop. I suspect
that the bug resides somewhere in the _preprocess function,
possibly due to an invalid path. May also have something to do with
the hpc cluster and slurm.